requirements.txt
asgiref==3.2.3
boto3==1.10.45
botocore==1.13.45
Django==3.0.1
django-environ==0.4.5
djangorestframework==3.11.0
docutils==0.15.2
jmespath==0.9.4
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
